+++ src/containerapp/HISTORY.rst
@@ -8,6 +8,7 @@ Release History
 * Add 'az containerapp env update' to update managed environment properties
 * Add custom domains support to 'az containerapp env create' and 'az containerapp env update'
 * Add tcp support for ingress transport and scale rules
+* `az containerapp up/github-action add`: Retrieve workflow file name from github actions API
 * 'az containerapp create/update': validate revision suffixes
 
 0.3.11